# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "root"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contatos": {
      "get": {
        "tags": [
          "Contatos"
        ],
        "description": "Buscar todos sem parametros ou passar parametros para filtrar",
        "parameters": [
          {
            "name": "idcontato",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Contatos"
        ],
        "description": "Cadastrar um contato",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codigo": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                },
                "telefone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "put": {
        "tags": [
          "Contatos"
        ],
        "description": "Atualizar um contato",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "codigo": {
                  "example": "any"
                },
                "nome": {
                  "example": "any"
                },
                "endereco": {
                  "example": "any"
                },
                "telefone": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/contatos/{idcontato}": {
      "get": {
        "tags": [
          "Contatos"
        ],
        "description": "Buscar por ID do contato",
        "parameters": [
          {
            "name": "idcontato",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contatos/{codigo}": {
      "delete": {
        "tags": [
          "Contatos"
        ],
        "description": "Excluir um contato informando a ID",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}